Rename package to match terraform.
All of our current lambdas are under uk.gov.nationalarchives as a package
so I think we'll stick with that for now. This matches what is in the
terraform.
